HRBOT/requirements.txt
llama-index
pypdf
pdfplumber
python-dotenv
ipykernel
google-generativeai>=0.3.0
chromadb
fastapi
uvicorn[standard]
sentence-transformers
pydantic
django-cors-headers
